Use AABB Acceleration Struct and Ray Queries
Replace the linear walk of a list of spheres in the rendering GLSL
with a walk of the BVH using accelerated ray queries.
This is currently actually slower than the version using the linear
scan of spheres but hopefully that is just because the scene is so
tiny and the queries will pull ahead when we draw more complex ones.