Added GLM to Thirdparty at Release 1.0.1
There is no use of the library in this commit.
The intention is to have it available for prototyping while equivalent
features are added to the minimal krust-gm and for comparisons in
krust-gm's unit tests. Of course it may turn out to be too useful to not
use more widely and the compile time and debugability hits from its
complexity and template use may not be as bad as I fear.